[display][coordinator] Move hotplug processing to client dispatcher.
This is a step towards simplifying the threading model for data
structures such as DisplayInfo, DisplayConfig, Image, and Layer.

This change may seem like a step backwards in responsitivity, as it will
cause the client dispatcher to become unresponsive while a display's
EDID is processed, which may entail multiple configuration checks. This
is not an issue for the following reasons.

1. In the medium term, all supported display hardware lacks EDID, so
   this change will not cause a regression.

2. In the long term, we're planning to move EDID processing to a
   dedicated panel driver.

Bug: 336394440
Test: fx test //src/graphics/display
Change-Id: If26f0baf9cf749ef2985900c0e003f1c3b97f816
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210125
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>