[kernel][vm] VmEnumerator is given Guard
Extends the EnumerateChildren method to pause enumeration during
callbacks, allowing it to pass the Guard in. This allows the callbacks
to temporarily drop the lock if desired.

While changing the interface to support the Guard parameter, the return
type is also changed to a zx_status_t to be more consistent with other
callback iteration patterns.

Change-Id: I50c8fb941221c77e71c2a05c6638637c9c9d1284
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1208171
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>