[virtio][pmem] Add basic virtio-pmem driver
This adds a driver for the Virtio PMEM device which is a mechanism for
sharing a region of "persisted" memory with a guest VM. This driver
reads the configuration of the device and creates a VMO representing the
persisted memory region. The VMO is not exposed outside the driver and
the "flush" operation is not implemented.

Change-Id: Iaaf7e5aac9b02c4c26e0b8924d7df4d7e9ded223
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213504
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>