[kernel] Remove idle time assert from scheduler
This idle time assertion has been causing a lot of flakes in CI/CQ,
so remove it. To handle the case where the actual_runtime is smaller
than the idle time (which is the case the assertion tested for),
bound the active_procesor_time_ns to 0. In other words, if the delta
would result in a negative number, clamp it to zero.

This workaround will be removed when the attached bug is resolved.

Bug: 379576294
Change-Id: Id13570f2c6de5117a4a1345ef77f8ce264ee5917
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215427
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>