[fx][rbe] Check RUNT network on corp machines.
In order to detect early a very annoying issue that leads
to failed remote builds, make `fx rbe auth` also verify that
the machine, if on a corp domain, is not connected to the
RUNT network.

This is done by calling the check-runt-script introduced
by tqr/948730, if avaiable in the current checkout.

In case of error, this willl print a message pointing to a
bug comment providing a temporary solution.

+ Remove a few shellcheck warnings.

Bug: 399096925
Change-Id: Idcd1239643f240d2436777b42dcd39f33eb04c3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213367
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: David Turner <digit@google.com>