[test_pilot] add TestConfigBuilder
This CL adds TestConfigBuilder to test_pilot. The builder implements
most of the functionality required to collect tests parameters and
assemble them into a test configuration. It leaves out the specific
set of parameters that test_pilot must understand, leaving that to
another struct.

This CL also add the Logger trait and a few implementations of it.
Loggers are used to debug the collection of test parameters.

Test: new unit tests

Bug: 327640651
Change-Id: Id6b7d76cae0cb902c282b539e875b18dd16f0248
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200685
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>