[zxdump] Handle ZX_INFO_PROCESS_MAPS with missing ZX_INFO_PROCESS_VMOS entry
This makes the zxdump memory dump logic more robust to getting a
ZX_INFO_PROCESS_VMOS sample that doesn't cover all the KOIDs that
appeared in the ZX_INFO_PROCESS_MAPS sample.  Heretofore this
would manifest as an assertion failure in a failed std::map::at
lookup.  Now it will just always elide the segment in question
from the dump as it would for a physical VMO or other "not safe
to attempt to dump" case.

It's suspected that this anomalous case is arising lately due to
a recent kernel regression.  But regardless, it's possible in two
ways for such an anomaly to arise without any bugs:
 1. Unsynchronized samples, either when dumping without
    suspending threads or when something outside the process is
    making mapping changes through a VMAR handle.
 2. Non-VMO mappings for iob and (soon) clock.

Bug: 399510219
Change-Id: I1a220c46d5292b535f79a93b5fb6078efd34ee45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216264
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>