[tee] Introduce ErrorWithSize
This is a simple abstraction packaging an error with size, useful for
operations that return a required minimum output size in the case of
Error::ShortBuffer.

While we're at it we revisit the related decision to have structured the
inner get_object_buffer_attribute() to return the written subslice of
the provided output buffer parameter in the case of success. While it's
the opinion of the author that returning the size of what was written
and leaving the synthesis of the written subslice as a boilerplate-y
exercise for the user is a particularly C-flavored API, that is the
flavor of API that we're ultimately implementing. (Plus the Rust
standard library seems to have capitalized a bit on this with the
definition of the io::Write trait.) So we update the success parameter
to just be a usize of how much was written, which will also ease the
implementation of several API bits to come.

Bug: 360942581
Change-Id: I3f53c5125aafe3b9918a74d39260ea3df343c8fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215425
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>