[gigaboot] shrink fastboot buffer to 1GiB
We previously allocated a 4GiB buffer for fastboot because we weren't
supporting sparse flashing, so we needed to be able to fit the multi-GiB
FVM image entirely in the buffer.

Now that we support sparse flashing we don't need such a large buffer,
and it was causing the emulator to fail to boot into fastboot since the
default emulator config only has 4GiB total.

Bug: b/399567210
Change-Id: I6fa135566bee920be5e2953ef88009b594dd10a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216591
Reviewed-by: Sergii Parubochyi <sergiip@google.com>
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>