[fdomain] Wake up client task when there are handles to close
We try to batch the protocol messages that ask an FDomain to close a
handle to save us round trips, but we weren't flushing the batches soon
enough so things that depend on observing a handle closure might break.

A test for a subsequent patch exposed this error, which is why this
patch has no test itself.

Change-Id: I47819ec3dca8134830766eff8f0218382246633f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188196
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>