[audio][drivers][aml-g12] Fix powered down state changes
When the hardware has been turned OFF, it may not be
safe to change its configuration via MMIO writes.
We now make sure that all MMIO writes are made with
the HW in an ON power state.

Also, we turn ON the HW on all new ring buffers since
we specify that the power state by default is ON
(The sdk/fidl/fuchsia.hardware.audio/ring_buffer.fidl
SetActiveChannels spec states that
"By default all channels are active").

Bug: 394673003
Test: fx test -o aml-g12-composite-test
Test: Trace power temporarily turning ON when writing to registers

Change-Id: I2942027d4a74ae6bb78a4487341a2f58465074de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1212945
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>