[starnix] Only return serialized bytes for SignalInfo to signals
For signals, one need to put the signalinfo struct associated with the
right ISA on the stack. This adds a method that always return an array
of the size of the 64 bit signal info, but filled with either the 32 or
64 bit version depending on the ISA.
This is used for putting the info on the stack, other callers should
always call write.

Bug: b/395946855
Change-Id: Iad70aea24fb7478d14b9a5ce42b122b9d1dec080
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216707
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>