[ftdi-i2c] Don't forward metadata
Make the ftdi-i2c driver add its metadata when the `FtdiI2c` class is
constructed instead of when `DdkInit()` is called.

Originally, when the class was constructed it would create a device and
forward its metadata. Once `DdkInit()` was called it would replace the
forwarded metadata with new metadata. This made it pointless to forward
the metadata in the first place.

Change-Id: I8fd92e970175b609df625a4b420859a6114b27d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215165
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>