[libvulkan] Use service instead of files for testing DeviceFs
The callback Vulkan-Loader currently provides for Fuchsia uses the
`fdio_service_connect` family of functions, which should not be used for
opening files. This callback is used to connect to the loader's
DeviceFs, which is intended to connect to a specific GPU device.

Implementations are only supposed to connect to devices (services)
through the loader's DeviceFs, so this isn't a problem in practice.
However, we do happen to rely on this for our fake implementation which
tests the Vulkan loader.

Currently this works because we do not correctly enforce filesystem
rights with the fdio_service_connect functions due to many callers
relying on implicit rights. This test is one of the last remaining
callers of fdio_service_connect that use it for "connecting" to
files/directories. Once we migrate this test, we can actually begin to
enforce the io2 rights model correctly.

Bug: 376575307
Test: fx test libvulkan-test
Change-Id: Ib264c0fcdcb8d4516065d5a38922b140310cff47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213344
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Craig Stout <cstout@google.com>