[component] Ensure services are opened as directories
Ensure that services are opened as directories instead of as protocol
connectors. For consistency with the GetServiceDirectory function, we
use the same set of permissions to ensure callers can enumerate protocol
instances within a service.

This is a precursor for migrating fdio_service_connect to use the new
Open method, as when we do, it would result in no permissions being
available on service instances.

Bug: 324111518
Bug: 376575307
Change-Id: I0edb75f0e0766ffd511a494dc10c96e90aaba028
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211864
API-Review: David Gilhooley <dgilhooley@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>