[assembly] Add support for bootfs files/packages in bootstrap and lower configurations
The bootstrap, embeddable, and empty configurations have a need to add
files or packages to bootfs.  This is rare, but needed for tests or
these configurations where the package system isn't fully available yet.

This allows product assembly to add to those sets based on files and
packages from the product (as opposed to the platform) but only in
these restricted modes.

Bug: b/399480273
Change-Id: If804ce1c61a9b51f35f2d2fdf3216afd42d95601
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215544
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>