[fake-coordinator-connector] Rework package structure for tests.
This CL reduces the complexity of integration tests by switching to GN
templates that are or will be used by other parts of the display driver
stack.

This CL also clarifies the reasons for using the "vulkan" test type,
which is subtle and surprising to a reader who is not familiar with the
test types.

Bug: 396160358
Test: fx test //src/graphics/display/testing/fake-coordinator-connector --count 100
Change-Id: Ia7c76566f7ad1bd178b476422306ea6f19f6249e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216331
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>