[starnix] Refactor empty children gating for unlink operations
This change helps mitigate the lock ordering bug described in
b/387576826. In short, the rename operation can potentially move
DirEntry nodes around the tree hierarchy. Since lock ordering assumes
a strict parent-to-child lock order, moving nodes around the tree can
result in false-positives for the mutex tracing graph.

The mitigation applied here is to offload the gating of non-empty
directories to the filesystem itself. This is arguably where the check
should belong in the first place, and has the added benefit of removing
the need for the DirEntry algorithm to obtain a lock on the children
of the node which we are unlinking.

This change adds a unittest to cover the above scenario. I've also
renamed "child" in the algorithm to instead be "child_to_unlink," which
I hope makes the code more readable.

Bug: b/387576826
Test: Locally validated, added a unittest, and CQ.
Change-Id: Id66e13d7c0100202465f9c2ba30157d4471c73a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206709
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>