[adr] Driver cmd timeouts
Institute timeouts for all* FIDL commands to Composite
drivers, so that ADR does not hang in the case of
unresponsive or late-responding drivers. Instead, these
drivers are marked as unhealthy and ultimately removed.

"All" FIDL commands entails the methods in the Composite,
Health, SignalProcessing and RingBuffer protocols
(20 methods in all), with the following clarifications:

* GetProperties(Composite), GetTopologies, GetElements,
  GetRingBufferFormats, GetDaiFormats, as well as the
  initial GetHealthState call are checked only in the
  aggregate, with a single longer timeout that brackets
  the entire the device initialization process. These
  methods are not called after initialization; during
  initialization we invoke them concurrently rather than
  serially, so guarding each method individually would
  unnecessarily complicate the driver command timeout
  mechanism without benefit. Subsequent GetHealthState
  calls ARE defended individually; these are in fact
  expressly used to surface timeouts during unittests.

* CreateRingBuffer, GetProperties(RingBuffer), GetVmo and
  the initial WatchDelayInfo call are checked only in the
  aggregate, with a single longer timeout that brackets the
  entire RingBuffer creation process. These methods are also
  called as a complete set in overlapped fashion, similar to
  the initialization-oriented methods mentioned above.

* "Hanging-get" methods WatchTopology, WatchElementState,
  WatchClockRecoveryPositionInfo and subsequent
  WatchDelayInfo calls are not defended by a driver timeout,
  since they are permitted to pend indefinitely before
  completing.

* The other 8 methods (SetTopology, SetElementState, Reset,
  SetDaiFormat, SetActiveChannels, Start, Stop and
  subsequent GetHealthState calls) are guarded by individual
  driver timeouts upon each invocation.

When a driver command response is not received within the
timeout interval (generally 2 sec, or 4 sec for the two
aforementioned "meta-commands"), the device's command state
is considered "overdue". If the response is later received,
then we log details but attempt to _recover_ the device to
normal operation. However, if the response has not yet been
received by the time ANOTHER command is issued, we give up
and consider the device "unresponsive"; it is first moved to
an Error state, then removed from the audio device registry.

Test: 10 new unittests, covering device initialization,
      RingBuffer creation and the 8 methods mentioned above.

fx test -dr audio_services_device_registry_warning_unittests

Change-Id: I6ae1f9a2e8111bfdf72b488f68ff4655f71aba9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214744
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>