[display][fidl][banjo] Clarify engine / coordinator API docs.
This CL updates the comments on most calls in the
fuchsia.hardware.display.engine/Engine protocol. The main goal is to
clarify the API contracts, with an emphasis on preconditions and
postconditions.

This CL also removes duplicate comments from the Banjo version of the
protocol above. The comments are replaced with references to the
equivalent calls in the FIDL protocol.

Bug: 316631158
Test: fx build (no code or API changes)
Change-Id: If72df3369cd7ff7d4ab53464d64aa97153139274
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202144
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
API-Review: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>