[idk][fidl] Unstable atoms only allowed in partner category
Require that FIDL libraries are stable unless they are in the "partner"
category by adding enforcement to `fidl()` similar to that already in
`sdk_source_set()` and `sdk_atom()`.

Also require `stable` to be specified when `sdk_category` is specified
and add it to the one library that didn't specify it.

In addition, simplify the logic enforcing stability in other categories
in `sdk_atom()`.

Bug: 397525399
Change-Id: I2650e71f653c0fb078dbd83f1c701531b3579892
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210320
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>