[kernel][vm] Rewrite PhysicalPageProvider
Changes the design of the PhysicalPageProvider to directly return all
the loaned pages to the PMM, instead of relying on the PMM to clean
them up via pmm_delete_lender.

As a consequence of this change the coordination between the
PhysicalPageProvider and the page requests can be improved to remove
some existing race conditions that meant it was necessary to avoid
canceling active page requests.

Change-Id: Iffb9fbfab2eed87f331daf0483f5e0a75b79252b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174214
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>