[kernel][vm] Update loaned state accessors
With the removal of pmm_delete_lender and other changes to the loaned
pages interface the access rules of loaned and loaned cancelled flags
can be more strictly defined.

Due to the need to pack into a single byte an atomic still needs to be
used, but logically manipulations of a particular flag are externally
coordinated and hence all reads and writes can at least be of a
relaxed memory ordering.

Change-Id: I08460ee09ed64e39d362a0c856253bf1fec68676
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174215
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>