[build] Add variant-libprefix conf to profile-use
This patch adds variant-libprefix config to profile-use
variant to ensure that the shared libraries are installed
into the appropriate path when built with that variant.

It also adds in nstrumented tag to the profile-use variant
because the special handling of shared library paths is
done for the variants based on the instrumented tag.

Bug: 335919004
Change-Id: Ia5b228e2b9f91b89de5f67b72201069601aff7b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217126
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gulfem Savrun Yeniceri <gulfem@google.com>
Reviewed-by: David Turner <digit@google.com>