[sys] Ensure `OpenNamedServiceAt` uses Open
Ensure that services are opened as directories instead of as protocol
connectors. This was already done for `OpenNamedServiceIn` but was
missed for `OpenNamedServiceAt`.

This is a precursor for migrating fdio_service_connect to use the new
Open method, as when we do, it would result in no permissions being
available on service instances.

Bug: 324111518
Bug: 376575307
Test: fx test //sdk/lib/sys
Change-Id: I59777acac2669141e5532c2182a6f44e7654a55e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217584
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>