[ffx][build] Remove sdk usage from cli-goldens
This reworks the cli-goldens tests to use ffx and the
subtool manifest instead of the SDK or the idk_atom_collection.

Also corrected metadata about ffx auth.

This improves the build performance since it no longer has to wait
for the IDK or SDK to be built.

Bug: 392136182
Change-Id: I5d62b813b04b82a1770bbb1140929ae2fd04d528
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216846
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>