[ffx][build] Remove sdk usage from mw-goldens
This reworks the mw-goldens tests to use ffx and the
subtool manifest instead of the SDK or the idk_atom_collection.

Also removes some unnecessary `async` function types.

And deletes a mis-located cli-golden file.

The change to the metrics code is to avoid loading the SDK to get the
SDK version when the analytics are disabled.

This improves the build performance since it no longer has to wait
for the IDK or SDK to be built.

Bug: 392136182
Change-Id: I648d1a58e17b4c76541795f8986f9d47ad4b8818
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217963
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>