[clidoc][build] Removing sdk-root option
This removes the use if the incorrectly named option "sdk-root"
this ends up being the build directory.

CLIDoc now uses a host tool atom collection to enumerate the tools
in included in the IDK, so there is no need for the IDK, or parts of
it, to be generated before calling clidoc.

Bug: 392136182
Change-Id: Iccb0b3b2e3804ce4563e3801acdd35d434572fce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216998
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>