[netstack3] Propagate more ICMP errors to TCP sockets
Currently there are a number of ICMP error codes that are not propagated
to TCP sockets, either as hard errors (which close the connection) or
soft errors (which don't close the connection but which can be retrieved
via SO_ERROR). Align the error propagation behavior with Linux, and add
integration test coverage for observing soft errors on established
connections, in addition to the existing coverage for hard errors on
attempted connections.

Bug: 42052672
Test: netstack3-core-tcp-test
      netstack-socket-integration-test
Change-Id: Iba0aff4bf319d21477cae264cdc0b4b62b74a154
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216279
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>