[kernel][arm64] Delay boot map init
This is a reland on http://fxrev.dev/1212556, which caused a crash in
start-up on sorrel. Previously the call was delayed until
arm64_mmu_early_init(), though it's unclear why that was an issue for
sorrel alone. In any case, it's clear that this call needs to be moved
up to before the periphmap is initialized (which is the last bit of code
directly dependent on the boot map)

We push back arm64_boot_map_init() to  platform_early_init(), just
before periphmap initialization. The boot map logic will go away
entirely soon, but until then there is no longer a need to call this in
assembly (as we're no longer setting up mappings in that context).

Bug: 42164859
Change-Id: I42f80479a2b6c7ae2684635a376152074520259b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1218005
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>