[starnix][cgroup] Store cgroup v2 fs on Expando.
Unlike v1, cgroup v2 has only single hierarchy. Moved to store `Cgroup2Fs`
on Expando so that it can be reused. `CgroupV2Fs` no longer keeps its
own `CgroupRoot` as it now lives on the Kernel.

Made `ScopedTempDir` movable but not copyable, which fixes the problem
where a new temp directory is created when it is implicitly copied (e.g.
when a `vector<ScopedTempDir>` is resized).

Bug: 389748287
Change-Id: I30d0925f987d2946963f1e69bbb037eed582e775
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215565
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: David Song <wintermelons@google.com>