[ffxutil] support ffx --strict emu start
This requires returning the "ffxInvoker", since the caller wants to
be able to both run the command, _and_ get the *exec.Cmd.

In strict, we require the emulator instance dir to be set, so we
now also add "emu.instance_dir" to the configs. And "ffx emu start"
specifically also requires "ssh.pub" to be set, so EmuStart() uses the
setConfigs() method to pass the additional config to the command.

Bug: 391949505
Change-Id: Idf19c806cd7ecdba306cdcb079cbff8c13ee2c30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217630
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>