[tee] Ensure tests open directories with permissions
Ensure the /ta directory and subdirectories are opened with permissions,
rather than "connecting" to them. This is a precursor for migrating
fdio_service_connect to the new Open method, as currently it uses the
DeprecatedOpen method which adds these rights implicitly. We wish to
restrict these rights and start enforcing the new io2 rights model,
which requires at least PERM_READABLE to enumerate directories.

Bug: 324111518
Bug: 376575307
Test: fx test //src/tee/manager
Change-Id: Ife3d5f0e703704e372044dd7f352430db377f009
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1219224
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>