[fxfs] Provide keys to create_object_with_id
Rather than having ObjectStore::create_object_with_id implicitly create
keys, require the caller to create them.

This is intended to make it easier for clients to handle the case where
key creation fails due to a Crypt error, for example during store
flushing when a child store's Crypt instance is unresponsive (see linked
bug).

For most callers which go via ObjectStore::create_object, this is a NOP,
as I just moved the key creation logic into that method.

A new ObjectStore::creat_object_with_key method is defined which must be
used for keys which are managed by a different store (e.g. the layer
files of an encrypted store, which live in the root store but are
encrypted with the store's key).

Bug: 399171573
Fixed: 364428824
Change-Id: I6717f80f2edd1428eedab7bbaa486d225a6a366c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1219975
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>