[kernel][vm] Return locked clone
When creating a clone from a VmCowPages returns the currently held lock
for the clone, allowing the caller to complete any setup under this
single lock acquisition.

This causes no change in behavior as with the hierarchy lock the parent
lock, which is already held, locks the child as well.

Bug: 338300943
Change-Id: Ife317789ad7dc0c1372a55eeca41abcfb25ead93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1218447
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>