[assembly][timekeeper] Firm up use_always_on_counter
We use the always on counter to emulate an RTC chip on platforms that do
not use the classic PC-like RTC circuitry.  But, for this to work we
need two ingredients: the counter itself (a board feature) and a
decision to use it (a platform config option).

This change makes the final selection be a logical AND of the two
settings.  This allows us to introduce the feature tag
`fuchsia::always_on_counter` without changing anything until
the platform bit is turned on.

This is not strictly part of b/388562148, but is part of the changes
needed for some platforms, so I'm attaching it to that bug.

Bug: 388562148
Change-Id: I0af538497af5c82d9ec45acb902cef361a0014e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220268
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>