[starnix] make memfd a tmpfile on a private tmpfs
This ensures that memfd is properly labeled for SEStarnix: we observed
that memfd is labeled as if it was a file create on tmpfs.

As supporting evidence for this change, consider the freshly-added
memfd.cc test and the policy change in
https://android-review.git.corp.google.com/c/platform/system/sepolicy/+/889855.

To get the right name in /proc/<pid>/fd, we fix the path-returning code
to add a `(deleted)` tag at the end of unreachable and deleted paths.

To allow testing, I change the base composite policy to give a specific label to
tmpfs.

Bug: 398816545
Change-Id: Ie5d1728929bb4f9a148fb1c462f9a2a667318247
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214904
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Ambre Williams <ambre@google.com>