[component_manager] increase liveness check frequency
We're still seeing crashes due to component manager writing to a full
channel, but we're missing the log message that's supposed to give us a
breadcrumb on where to look.

Increase the frequency of this log message to every 500th write. This
will both make the log more noticeable, and ensure that the log message
is written shortly before the crash (the write limit of 3500 and a log
interval of 1000 meant it got written 500 writes before the crash).

Bug: 393176930
Change-Id: I157cd03347f8d902ddd2315eeb0ece4dd73b4d0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220965
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>