[idk][build] Build less for the in-tree IDK
Create separate IDK build graphs for the IDK used in-tree (by the Bazel
in-tree SDK and ffx tests) and the official IDK distributed to partners.
This decreases the build graph for most developers.

"//sdk:bazel_in_tree_idk" and "//sdk:sdk" now depend on the in-tree
graph. This avoids building unneeded targets when, for example,
assembling with the Bazel SDK.

"final_fuchsia_idk.exported" depends on the complete IDK graph, which is
now called "fuchsia".

The template named "idk_rules" implements the common logic for the two
graphs, which provide the specific targets to include when instantiating
the template.

Remove ":sdk_source_set_list" from "//sdk" since it is already a direct
dependency of "//:api" and would otherwise pull in the full IDK.

The same set of host tools are included in a default build, but tools
not needed as part of the SDK build are not included in the in-tree IDK.
This allows these tools to be built in parallel with targets that depend
on the IDK and SDK.

Also partially fix the dependency checking that triggers sub-build
rebuilds by adding a non-SDK deps in the main build on targets that are
produced for the IDK only in sub-builds. Packages are the remaining gap.

Bug: 393154095
Change-Id: I6375869c3d1a2be082efe5fe0b57ea73ce1c31cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198649
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>