[assembly] Assembly containers do not write paths with colons
Bazel does not like paths with colons, therefore before we write a
file to a container we convert any colons to a directory separator.

For example, if previously we would have written to:
  container/path/inside/with::colon

We now write to:
  container/path/inside/with/colon

As seen from the example, we also properly handle multiple colons in
a row.

Change-Id: I418def7eeb1803e1de71c1d3bde6e8960e869be0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220845
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>