[fbl] Use placement new instead of reinterpret_cast
I was experimenting with CFI in the kernel and CFI complained about this
cast between unrelated types. In the ctor for fbl::InlineArray, we
reinterpret_cast from a char array to a WaitSignalObserver pointer.
These don't share any type hierarchy. Placement new should be the
correct way to do this.

Bug: 345527716
Change-Id: I0e7d39df4649f8eccbf20ec797c8b9aeaabe3f69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198565
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>