[tee] Hoist "ta" collection up to peer of ta-manager
This moves the "ta" collection to be a peer of the ta-manager component.
This matches the test setup. This makes it easier to set up test realms
as the capability for controlling the collection also controls
subpackage resolution. Now we can use the same ta-manager component in
production and tests and supply test TAs as subpackages in test realms.

This also moves the boilerplate capabilities and routing for hosting the
TA manager into a shard that we can use in tee.cml and (most) tests.

Change-Id: I43764f8d3fdc97088821611f8a835d98c003400d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220285
Reviewed-by: Eric Stone <ecstone@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>