[console] Use service capabilities
The topological path suffix was removed as the only user of this
API specifies a virtio-console and it will only ever spawn a single
virtio-console. At that point, we just know the console is from a driver
rather than the kernel, and that's sufficient for us to just connect to
the first service instance to enumerate, so we switch to providing a bool
instead.

Change-Id: I76bdd8d204c312536ba82ce2cecdb7e8e904561c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216996
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Garratt Gallagher <garratt@google.com>