[kernel][iommu] New Map/Unmap IOMMU API
Changes the Map/Unmap design of the Iommu interface to use an opaque
mapping token, instead of operating directly on device address ranges.
This allows for separating the mapping of a range, with querying the
device contiguous portions of it. The existing IOMMU implementations
are able to support this change without actually needing to allocate or
store any additional metadata.

Bug: 399697584
Change-Id: Idd8b2d3fad2b924aa7c21cab6a63573591c8b3a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216904
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>