[kernel][bti] Minimize address buffering
Instead of buffering the entire set of addresses in a sys_bti_pin
operation, uses a helper to allow for buffering small amounts with
automatic flushing. This avoids any temporary heap allocations for
storing the addresses, as this smaller buffer can be stack allocated.

Bug: 399697584
Change-Id: I14a17c7c0e99b5d032c6826b4cbc019a3f54fd33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216907
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>