[dash] Update 'k' builtin to log an error if it can't send a debug command
Print an error message to stderr if this built-in is unable to send
a debug command to the kernel. When running this command from a non-serial
shell, such as a ssh shell, the command produces no direct output and so
failures are easy to miss.

Change-Id: Id04fafe6649be283e1a82e82cbd52c4a1027ee67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221347
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Justin Mattson <jmatt@google.com>