[kernel][vm] PhysicalPageProvider waits
Instead of expecting the event in WaitOnEvent to already be signaled
after processing requests block with a timeout to allow for greater
robustness against known, and unknown, racy interractions.

With the event now blocking the lock to serialize requests handling is
redundant and removed.

Bug: 400838270
Change-Id: I4f25de9d57c858c8748bbd9b4ef5eba885553333
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221486
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>