[power] Balance system-activity-governor tasks
The optimization of the SAG has greatly reduced the execution time at a
client side for calling TakeWakeLease. But this resulted in many tasks
scheduled but not completed at the end of the benchmark test for
TakeWakeLease, and happen to affect the result of ToggleLease because
Power Broker was still busy handling the requests from the tasks on the
SAG side.

To avoid this, a semaphore is added to control the flow so that tasks
can get executed. The test is also changed to hold to a lease so that
the optimization of the lease reuse can be tested. A different test for
the original behavior may be added later.

Bug: b/395214643
Test: refer to the bug.
Change-Id: Ie3091d711fa7957658d8c686f8fc158fe3093b57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217325
Reviewed-by: Michael Brunson <mbrunson@google.com>
Commit-Queue: Hai Bi <bihai@google.com>